functions/requirements.txt
firebase_functions~=0.1.0
Flask==3.0.3
opentelemetry-sdk==1.25.0
